refactor(rubrics): Group rubrics DTOs
