--- src/rubrics/domain/dtos/add_criteria_to_objective_dto.go
@@ -1,8 +1,0 @@
-package dtos
-
-type AddCriteriaToObjectiveDTO struct {
-	TeacherUUID         string
-	ObjectiveUUID       string
-	CriteriaDescription string
-	CriteriaWeight      float64
-}
