--- src/rubrics/domain/dtos/add_objective_to_rubric_dto.go
@@ -1,7 +1,0 @@
-package dtos
-
-type AddObjectiveToRubricDTO struct {
-	TeacherUUID          string
-	RubricUUID           string
-	ObjectiveDescription string
-}
