--- src/rubrics/domain/dtos/create_rubric_dto.go
@@ -1,6 +1,0 @@
-package dtos
-
-type CreateRubricDTO struct {
-	TeacherUUID string
-	Name        string
-}
