--- src/rubrics/domain/dtos/created_rubrics_dto.go
@@ -1,7 +1,0 @@
-package dtos
-
-type CreatedRubricDTO struct {
-	UUID        string `json:"uuid"`
-	TeacherUUID string `json:"-"`
-	Name        string `json:"name"`
-}
