+++ src/rubrics/domain/dtos/criteria_dtos.go
@@ -1,5 +1,12 @@
 package dtos
 
+type AddCriteriaToObjectiveDTO struct {
+	TeacherUUID         string
+	ObjectiveUUID       string
+	CriteriaDescription string
+	CriteriaWeight      float64
+}
+
 type UpdateCriteriaDTO struct {
 	TeacherUUID         string
 	CriteriaUUID        string
