--- src/rubrics/domain/dtos/delete_objective_dto.go
@@ -1,6 +1,0 @@
-package dtos
-
-type DeleteObjectiveDTO struct {
-	TeacherUUID   string `json:"teacher_uuid" binding:"required"`
-	ObjectiveUUID string `json:"objective_uuid" binding:"required"`
-}
