--- src/rubrics/domain/dtos/get_rubric_by_uuid_dto.go
@@ -1,6 +1,0 @@
-package dtos
-
-type GetRubricDto struct {
-	TeacherUUID string
-	RubricUUID  string
-}
