--- src/rubrics/domain/dtos/update_objective_dto.go
@@ -1,7 +1,0 @@
-package dtos
-
-type UpdateObjectiveDTO struct {
-	TeacherUUID        string
-	ObjectiveUUID      string
-	UpdatedDescription string
-}
