--- src/rubrics/domain/dtos/update_rubric_name_dto.go
@@ -1,7 +1,0 @@
-package dtos
-
-type UpdateRubricNameDTO struct {
-	TeacherUUID string
-	RubricUUID  string
-	Name        string
-}
